+++ R/micromake.R
@@ -9,10 +9,15 @@
 #'
 #' @param x A px object.
 #' @param out_dir Directory to save px files in.
+#' @param keyword_values Optional. A data frame with column 'variable' and one
+#' or more of: 'contents', 'title', 'description', and 'matrix'. The columns
+#' with be added as keywords to the table for each non-HEADING varabe that
+#' match the 'variable' column. It probably work for other keywords as well.
+#' Currently doesn't support multiple languages.
 #'
 #' @return Nothing
 #' @export
-micromake <- function(x, out_dir = NULL) {
+micromake <- function(x, out_dir = NULL, keyword_values = NULL) {
   validate_micromake_arguments(x, out_dir)
 
   print_out_dir <- is.null(out_dir)
@@ -29,6 +34,14 @@ micromake <- function(x, out_dir = NULL) {
     x %>%
     stub(micro_vars)
 
+  if (! is.null(keyword_values)) {
+    keyword_values_long <-
+      keyword_values %>%
+      tidyr::pivot_longer(cols = setdiff(names(.), c("variable", "language")),
+                          names_to = "keyword_function"
+                          )
+  }
+
   for (micro_var in micro_vars) {
     new_data <-
       x$data %>%
@@ -39,7 +52,8 @@ micromake <- function(x, out_dir = NULL) {
 
     data_names <- names(new_data)
 
-    new_px(languages  = new_px$languages,
+    x_micro <-
+      new_px(languages  = new_px$languages,
            table1     = new_px$table1,
            table2     = new_px$table2,
            variables1 = dplyr::filter(new_px$variables1, `variable-code` %in% data_names),
@@ -49,8 +63,23 @@ micromake <- function(x, out_dir = NULL) {
            data       = new_data
            ) %>%
       fix_px() %>%
-      figures(figures_var) %>%
-      pxsave(path = file.path(out_dir, paste0('micro_', micro_var, '.px')))
+      figures(figures_var)
+
+    if(!is.null(keyword_values)) {
+      extra_keywords <-
+        keyword_values_long %>%
+        dplyr::filter(variable %in% micro_var)
+
+      for (i in 1:nrow(extra_keywords)) {
+        modifying_function <- get(extra_keywords$keyword_function[i])
+
+        x_micro <- modifying_function(x = x_micro, value = extra_keywords$value[i])
+      }
+    }
+
+    pxsave(x = x_micro,
+           path = file.path(out_dir, paste0(micro_var, '.px'))
+           )
   }
 
   if (print_out_dir) print(paste("Created px files in:", out_dir))